Move out State to its own module
